Scenarios
I've written the guide section for Scenarios (and removed the Events.md file, it was redundant, since I did not see any need to separate the two). No images, but should be thorough enough for the time being, just so that it is written somewhere how to actually use scenarios.